Use Steam libraryfolders.vdf on Linux/Mac
Prompted by me being annoyed that my system couldn't detect these because I have my games on a different drive.

I went ahead and tried to unify the shared parts of the code to try and prevent this happening again in the future. The platform code is now only handling finding the path to the file.